Move details to seperate folder
