+++ src/details/Details.css
