+++ src/details/Details.js
